Вынес приведение за скобки

+++ src/main/kotlin/Main.kt
@@ -3,7 +3,7 @@ fun main() {
     val secondStudent: Byte = 4
     val thirdStudent: Byte = 3
     val fourthStudent: Byte = 5
-    val sum: Double = (firstStudent.toDouble() + secondStudent.toDouble() +
-            thirdStudent.toDouble() + fourthStudent.toDouble()) / 4
+    val sum: Double = (firstStudent + secondStudent +
+            thirdStudent + fourthStudent).toDouble() / 4
     println(sum)
 }
